protect private routes with authMiddleware

+++ backend/controllers/contactController.js
@@ -9,7 +9,13 @@ const controller = {}
  * @access Private
  */
 controller.getContacts = asyncHandler(async (req,res) => {
-  const contacts = await Contact.find();
+  // Check if user exists
+  if (!req.user) {
+    res.status(401);
+    throw new Error('User not found');
+  }
+
+  const contacts = await Contact.find({user: req.user.id});
   res.status(200).json(contacts);
 });
 
@@ -32,9 +38,15 @@ controller.setContact = asyncHandler(async (req,res) => {
     throw new Error(errors.array().map(error => error.msg).join('\n'))
   }
 
+  // Check if user exists
+  if (!req.user) {
+    res.status(401);
+    throw new Error('User not found');
+  }
+
   const {name, email, phone, type} = req.body;
 
-  const contact = await Contact.create({name, email, phone, type})
+  const contact = await Contact.create({user: req.user.id ,name, email, phone, type})
   res.status(201).json(contact);
 });
 
@@ -50,6 +62,21 @@ controller.updateContact = asyncHandler(async (req,res) => {
   if (!contact) {
     res.status(400);
     throw new Error('Contact not found')
+  }
+
+  // Check if user exists
+  if (!req.user) {
+    res.status(401);
+    throw new Error('User not found');
+  }
+
+  // console.log(contact.user, typeof contact.user)
+  // console.log(req.user.id, typeof req.user.id);
+
+  // Make sure the logged in user matches the contact user
+  if (contact.user.toString() !== req.user.id) {
+    res.status(401);
+    throw new Error('User not authorized');
   }
 
   const updatedContact = await Contact.findByIdAndUpdate(req.params.id, req.body, {new: true});
@@ -69,6 +96,18 @@ controller.deleteContact = asyncHandler(async (req,res) => {
   if (!contact) {
     res.status(400);
     throw new Error('Contact not found')
+  }
+
+  // Check if user exists
+  if (!req.user) {
+    res.status(401);
+    throw new Error('User not found');
+  }
+
+  // Make sure the logged in user matches the contact user
+  if (contact.user.toString() !== req.user.id) {
+    res.status(401);
+    throw new Error('User not authorized');
   }
 
   await Contact.findByIdAndRemove(req.params.id);
